Throw OOME if Bitmap.nativeCreate fails
This matches the old behavior, prior to switching from Java allocated
pixel memory to native allocations (b/27762775). It also better matches
what has happened - we ran out of memory. (Better than the current
behavior - NullPointerException in the Java code, or an alternative
solution of returning null, which would likely result in NPEs in the
calling code.)

Change-Id: I3958ed1106ac94fb1d3f30e044b620d984875211